Fix: Always guarantee the correct memory length.
If start + len overflowed, the previous code would not catch it and
it would return an empty slice (as start would be > end).
This could lead to an oob access if the caller expected a slice with the
exact requested length.

This change ensures that if start + len overflows, an error is returned.